chenge route of LateralCard
